+++ src/sbs/controller.cpp
@@ -80,7 +80,7 @@ void DispatchController::CheckArrivals()
 
 	SBS_PROFILE("DispatchController::CheckArrivals");
 
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		Elevator *elev = sbs->GetElevator(Elevators[i].number);
 
@@ -88,7 +88,7 @@ void DispatchController::CheckArrivals()
 		if (Elevators[i].arrived == true)
 		{
 			//step through requests table and find a request that matches the arrived elevator
-			for (int j = 0; j < Routes.size(); j++)
+			for (size_t j = 0; j < Routes.size(); j++)
 			{
 				//get request's direction of travel
 				int direction_dd = 0;
@@ -125,7 +125,7 @@ void DispatchController::CheckArrivals()
 							bool result = false;
 
 							//dispatch elevator to destination floor for each registered assignment
-							for (int k = 0; k < Elevators[i].assigned_destination.size(); k++)
+							for (size_t k = 0; k < Elevators[i].assigned_destination.size(); k++)
 							{
 								if (Routes[j].destination_floor == Elevators[i].assigned_destination[k])
 								{
@@ -173,7 +173,7 @@ void DispatchController::RemoveRoute(Route &route)
 {
 	//remove a destination dispatch route
 
-	for (int i = 0; i < Routes.size(); i++)
+	for (size_t i = 0; i < Routes.size(); i++)
 	{
 		if (Routes[i].starting_floor == route.starting_floor && Routes[i].destination_floor == route.destination_floor)
 		{
@@ -249,7 +249,7 @@ bool DispatchController::RequestRoute(CallStation *station, int starting_floor, 
 	}
 
 	//use existing entry if found
-	for (int i = 0; i < Routes.size(); i++)
+	for (size_t i = 0; i < Routes.size(); i++)
 	{
 		if (Routes[i].starting_floor == starting_floor && Routes[i].destination_floor == destination_floor && Routes[i].requests < MaxPassengers)
 		{
@@ -343,7 +343,7 @@ bool DispatchController::CallElevator(CallStation *station, bool direction)
 		dir = -1;
 
 	//use existing entry if found
-	for (int i = 0; i < Routes.size(); i++)
+	for (size_t i = 0; i < Routes.size(); i++)
 	{
 		if (Routes[i].starting_floor == floor && Routes[i].direction == dir)
 		{
@@ -381,7 +381,7 @@ void DispatchController::ProcessRoutes()
 
 	SBS_PROFILE("DispatchController::ProcessRoutes");
 
-	for (int i = 0; i < Routes.size(); i++)
+	for (size_t i = 0; i < Routes.size(); i++)
 	{
 		if (Routes[i].processed == true)
 		{
@@ -493,7 +493,7 @@ void DispatchController::ProcessRoutes()
 			else
 			{
 				//update elevator's call information
-				for (int j = 0; j < Elevators.size(); j++)
+				for (size_t j = 0; j < Elevators.size(); j++)
 				{
 					if (Elevators[j].number == Routes[i].assigned_elevator)
 					{
@@ -540,7 +540,7 @@ void DispatchController::ProcessRoutes()
 		//update elevator's call information
 		if (Routes[i].destination == true)
 		{
-			for (int j = 0; j < Elevators.size(); j++)
+			for (size_t j = 0; j < Elevators.size(); j++)
 			{
 				if (Elevators[j].number == elevator->Number)
 				{
@@ -817,7 +817,7 @@ void DispatchController::ElevatorArrived(int number, int floor, bool direction)
 
 	SBS_PROFILE("DispatchController::ElevatorArrived");
 
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		//exit if elevator already arrived
 		if (Elevators[i].number == number && Elevators[i].arrived == true)
@@ -832,7 +832,7 @@ void DispatchController::ElevatorArrived(int number, int floor, bool direction)
 			Report("Elevator " + ToString(number) + " arrived at floor " + ToString(floor) + " (" + floorobj->ID + ")");
 
 			//only set values if arriving at starting floor
-			for (int j = 0; j < (int)Elevators[i].calls.size(); j++)
+			for (size_t j = 0; j < (int)Elevators[i].calls.size(); j++)
 			{
 				if (Elevators[i].calls[j].floor == floor && Elevators[i].calls[j].direction == direction)
 				{
@@ -847,7 +847,7 @@ void DispatchController::ElevatorArrived(int number, int floor, bool direction)
 			if (Elevators[i].assigned == true)
 			{
 				//unassign route from elevator in destination dispatch mode
-				for (int j = 0; j < Elevators[i].assigned_destination.size(); j++)
+				for (size_t j = 0; j < Elevators[i].assigned_destination.size(); j++)
 				{
 					if (Elevators[i].assigned_destination[j] == floor)
 					{
@@ -857,7 +857,7 @@ void DispatchController::ElevatorArrived(int number, int floor, bool direction)
 				}
 
 				//unassign route from elevator in standard mode
-				for (int j = 0; j < Elevators[i].assigned_directions.size(); j++)
+				for (size_t j = 0; j < Elevators[i].assigned_directions.size(); j++)
 				{
 					int direction2 = -1;
 					if (direction == true)
@@ -905,14 +905,14 @@ bool DispatchController::IsElevatorAssigned(int number, int destination_floor, i
 {
 	//return true if the specified elevator is assigned to the specified destination floor or direction
 
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		if (Elevators[i].number == number)
 		{
 			if (direction == 0)
 			{
 				//destination dispatch mode
-				for (int j = 0; j < Elevators[i].assigned_destination.size(); j++)
+				for (size_t j = 0; j < Elevators[i].assigned_destination.size(); j++)
 				{
 					if (Elevators[i].assigned_destination[j] == destination_floor && Elevators[i].assigned == true)
 						return true;
@@ -921,7 +921,7 @@ bool DispatchController::IsElevatorAssigned(int number, int destination_floor, i
 			else
 			{
 				//standard mode
-				for (int j = 0; j < Elevators[i].assigned_directions.size(); j++)
+				for (size_t j = 0; j < Elevators[i].assigned_directions.size(); j++)
 				{
 					if (Elevators[i].assigned_directions[j] == direction && Elevators[i].assigned == true)
 						return true;
@@ -938,7 +938,7 @@ bool DispatchController::IsElevatorAssignedToOther(int number, int destination_f
 	//return true if the specified elevator is assigned to a destination floor or direction other than
 	//the specified one, while also taking the serviced floor range into account if more than one elevator is assigned
 
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		if (Elevators[i].number == number)
 		{
@@ -974,7 +974,7 @@ void DispatchController::AssignElevator(int number, int destination_floor, int d
 {
 	//assign an elevator to the specified destination floor
 
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		if (Elevators[i].number == number)
 		{
@@ -988,7 +988,7 @@ void DispatchController::AssignElevator(int number, int destination_floor, int d
 			else //standard mode
 			{
 				//exit if elevator has already been assigned to the direction
-				for (int j = 0; j < Elevators[i].assigned_directions.size(); j++)
+				for (size_t j = 0; j < Elevators[i].assigned_directions.size(); j++)
 				{
 					if (Elevators[i].assigned_directions[j] == direction)
 					{
@@ -1009,7 +1009,7 @@ void DispatchController::RegisterCallStation(CallStation *station)
 	//register the specified call station
 
 	//exit if already registered
-	for (int i = 0; i < CallStations.size(); i++)
+	for (size_t i = 0; i < CallStations.size(); i++)
 	{
 		if (CallStations[i] == station)
 			return;
@@ -1053,7 +1053,7 @@ int DispatchController::GetElevatorArrived(int starting_floor, int destination_f
 	//return the number of the elevator that has arrived, for the specified route
 	//return 0 if no elevator has arrived yet
 
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		if (Elevators[i].arrived == true)
 		{
@@ -1069,7 +1069,7 @@ int DispatchController::GetElevatorArrivedStandard(int floor, bool direction)
 	//return the number of the elevator that has arrived, for the specified direction
 	//return 0 if no elevator has arrived yet
 
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		if (Elevators[i].arrived == true)
 		{
@@ -1155,7 +1155,7 @@ bool DispatchController::AtMaxRequests(int elevator, int destination_floor)
 
 	//get elevator index
 	int index = -1;
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		if (Elevators[i].number == elevator)
 		{
@@ -1165,9 +1165,9 @@ bool DispatchController::AtMaxRequests(int elevator, int destination_floor)
 	}
 
 	//return true if an elevator's assigned DD route has requests at the MaxPassengers value
-	for (int i = 0; i < Routes.size(); i++)
+	for (size_t i = 0; i < Routes.size(); i++)
 	{
-		for (int j = 0; j < Elevators[index].assigned_destination.size(); j++)
+		for (size_t j = 0; j < Elevators[index].assigned_destination.size(); j++)
 		{
 			if (Elevators[index].assigned_destination[j] == destination_floor && Elevators[index].assigned == true)
 			{
@@ -1183,7 +1183,7 @@ bool DispatchController::AtMaxRequests(int elevator, int destination_floor)
 void DispatchController::ResetArrival(int number)
 {
 	//reset arrival status
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		if (Elevators[i].number == number)
 		{
@@ -1268,7 +1268,7 @@ bool DispatchController::SameElevators(const std::vector<int> &elevators)
 	if (Elevators.size() != elevators.size())
 		return false;
 
-	for (int i = 0; i < (int)Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		if (Elevators[i].number != elevators[i])
 			return false;
@@ -1285,11 +1285,11 @@ bool DispatchController::GetCallStatus(int elevator, int floor, bool &up, bool &
 	down = false;
 	bool result = false;
 
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		if (Elevators[i].number == elevator)
 		{
-			for (int j = 0; j < (int)Elevators[i].calls.size(); j++)
+			for (size_t j = 0; j < Elevators[i].calls.size(); j++)
 			{
 				if (Elevators[i].calls[j].floor == floor)
 				{
@@ -1312,7 +1312,7 @@ bool DispatchController::ServesFloor(int floor)
 {
 	//return true if an associated elevator serves the specified floor
 
-	for (int i = 0; i < Elevators.size(); i++)
+	for (size_t i = 0; i < Elevators.size(); i++)
 	{
 		Elevator *elev = sbs->GetElevator(Elevators[i].number);
 		if (elev)
